requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic[email]==2.12.5
pydantic-settings==2.4.0
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.34
asyncpg==0.30.0
httpx==0.27.0
ortools==9.15.6755
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 개발/테스트
pytest==9.0.2
pytest-asyncio==1.3.0
